feat: cut over horseshoes to path engine
