+++ src/const.js
@@ -13,7 +13,6 @@ const DEFAULT_ZPOS = {
   circles: 300,
   horseshoes: 400,
   horseshoes_v2: 400,
-  horseshoes_v3: 400,
   lines: 500,
   hlines: 500,
   vlines: 500,
@@ -32,7 +31,6 @@ const DEFAULT_RENDER_INDEX = {
   circles: 300000,
   horseshoes: 400000,
   horseshoes_v2: 400000,
-  horseshoes_v3: 400000,
   lines: 500000,
   hlines: 500000,
   vlines: 500000,
