+++ src/horseshoe-geometry.js
@@ -1,12 +1,4 @@
-/* eslint-disable max-classes-per-file */
-
 import { clamp } from './frontend_mods/common/number/clamp.ts';
-import { PathValueMapper } from './path-ranges.js';
-
-/**
- * Conversion factor used by point projection helpers.
- */
-const DEG_TO_RAD = Math.PI / 180;
 
 /**
  * Cubic Hermite spline used by the original spline scale type.
@@ -303,253 +295,5 @@ export class GaugeScale {
     }
 
     return clamp((numericValue - this.min) / (this.max - this.min), 0, 1);
-  }
-}
-
-/**
- * Converts normalized scale positions into angles, transforms, and SVG coordinates.
- */
-export class GaugeGeometry {
-  /**
-   * Captures the current horseshoe scale branch and converts normalized ratios
-   * into the configured arc's transformed coordinate space.
-   */
-  constructor(config, scale, activeValue) {
-    this.cx = config.svg.xpos;
-    this.cy = config.svg.ypos;
-
-    this.radius = config.svg.radius;
-    this.tickmarksRadius = config.svg.tickmarks_radius;
-
-    this.arcDegrees = config.arc_degrees;
-    this.startAngle = config.start_angle;
-    this.endAngle = this.startAngle + this.arcDegrees;
-
-    this.rotation = Number(config.rotate ?? 0);
-    this.flip = config.flip ?? 'none';
-    this.groupConfig = config.group_config;
-    this.barMode = config.bar_mode;
-    this.absoluteSign = Number(activeValue) < 0 ? -1 : 1;
-
-    this.zeroRatio = config.zero_ratio;
-    this.zeroAngle = this.ratioToAngle(this.zeroRatio);
-
-    this.scale = scale;
-    this.valueMapper = new PathValueMapper({
-      scale,
-      barMode: this.barMode,
-      zeroRatio: this.zeroRatio,
-      stateMode: config.horseshoe_state.mode,
-      stateMap: config.state_map?.map,
-    }, activeValue);
-  }
-
-  /**
-   * Selects the signed scale branch used by absolute scale, background, tick,
-   * label, and gradient layers. Zero belongs to the positive branch.
-   *
-   * @param {number} value - Current signed entity or animated display value.
-   * @returns {boolean} Whether the active branch changed.
-   */
-  setActiveValue(value) {
-    const changed = this.valueMapper.setActiveValue(value);
-    this.absoluteSign = this.valueMapper.absoluteSign;
-
-    return changed;
-  }
-
-  /**
-   * Returns the signed source interval represented from the physical start to
-   * the end of an absolute horseshoe. A 0..max scale mirrors max for negatives.
-   *
-   * @returns {{start: number, end: number}} Active signed source interval.
-   */
-  getActiveSourceRange() {
-    return this.valueMapper.getActiveSourceRange();
-  }
-
-  /**
-   * Converts a positive visual magnitude into the signed value used for scale
-   * mapping and color-stop lookup on the active absolute branch.
-   *
-   * @param {number} magnitude - Positive distance from zero.
-   * @returns {number} Signed source value.
-   */
-  magnitudeToSourceValue(magnitude) {
-    return this.valueMapper.magnitudeToSourceValue(magnitude);
-  }
-
-  /** Returns the visible positive magnitude limit of the active branch. */
-  getActiveMagnitudeMax() {
-    return this.valueMapper.getActiveMagnitudeMax();
-  }
-
-  /**
-   * Selects signed color stops from the active branch and orders them from the
-   * physical horseshoe start toward its end.
-   *
-   * @param {Array<object>} colorStops - Normalized signed color stops.
-   * @returns {Array<object>} In-range color stops in visual arc order.
-   */
-  getActiveColorStops(colorStops) {
-    return this.valueMapper.getActiveColorStops(colorStops);
-  }
-
-  /**
-   * Returns rotation and flip flags used by renderers and labels.
-   *
-   * @returns {object} Transform context for child renderers.
-   */
-  getTransformContext() {
-    return {
-      rotation: this.rotation,
-      flipX: this.flip === 'x' || this.flip === 'both',
-      flipY: this.flip === 'y' || this.flip === 'both',
-    };
-  }
-
-  /**
-   * Builds the SVG rotate transform around the gauge center.
-   */
-  getRotateTransform() {
-    if (!this.rotation) {
-      return '';
-    }
-
-    return `rotate(${this.rotation} ${this.cx} ${this.cy})`;
-  }
-
-  /**
-   * Builds the SVG flip transform around the gauge center.
-   */
-  getScaleTransform() {
-    const transformContext = this.getTransformContext();
-
-    if (!transformContext.flipX && !transformContext.flipY) {
-      return '';
-    }
-
-    const scaleX = transformContext.flipX ? -1 : 1;
-    const scaleY = transformContext.flipY ? -1 : 1;
-
-    return `translate(${this.cx} ${this.cy}) scale(${scaleX} ${scaleY}) translate(${-this.cx} ${-this.cy})`;
-  }
-
-  /**
-   * Builds the optional group rotation around the group center.
-   */
-  getGroupRotateTransform() {
-    const groupRotation = Number(this.groupConfig?.rotate ?? this.groupConfig?.rotation ?? 0);
-
-    if (!groupRotation) {
-      return '';
-    }
-
-    return `rotate(${groupRotation} ${this.groupConfig.svg.xpos} ${this.groupConfig.svg.ypos})`;
-  }
-
-  /**
-   * Builds the optional group scale around the group center.
-   */
-  getGroupScaleTransform() {
-    if (!this.groupConfig?.scale) {
-      return '';
-    }
-
-    const scaleX = this.groupConfig.scale.x ?? this.groupConfig.scale;
-    const scaleY = this.groupConfig.scale.y ?? this.groupConfig.scale;
-
-    return `translate(${this.groupConfig.svg.xpos} ${this.groupConfig.svg.ypos}) scale(${scaleX} ${scaleY}) translate(${-this.groupConfig.svg.xpos} ${-this.groupConfig.svg.ypos})`;
-  }
-
-  /**
-   * Combines group transforms with rotation and flip transforms for the gauge group.
-   */
-  getGroupTransform() {
-    return [
-      this.getGroupRotateTransform(),
-      this.getGroupScaleTransform(),
-      this.getRotateTransform(),
-      this.getScaleTransform(),
-    ].filter(Boolean).join(' ');
-  }
-
-  /**
-   * Builds the inverse transform used to keep labels readable.
-   */
-  getInverseGroupTransform() {
-    const transformContext = this.getTransformContext();
-    const parts = [];
-
-    if (transformContext.flipX || transformContext.flipY) {
-      const scaleX = transformContext.flipX ? -1 : 1;
-      const scaleY = transformContext.flipY ? -1 : 1;
-
-      parts.push(`translate(${this.cx} ${this.cy})`);
-      parts.push(`scale(${scaleX} ${scaleY})`);
-      parts.push(`translate(${-this.cx} ${-this.cy})`);
-    }
-
-    if (this.rotation) {
-      parts.push(`rotate(${-this.rotation} ${this.cx} ${this.cy})`);
-    }
-
-    return parts.join(' ');
-  }
-
-  /**
-   * Converts a normalized scale position into an absolute arc angle.
-   */
-  ratioToAngle(ratio) {
-    return this.startAngle + ratio * this.arcDegrees;
-  }
-
-  /**
-   * Converts a scale value into the raw configured scale ratio.
-   *
-   * @param {number} value - Scale value to map.
-   * @returns {number} Normalized scale position without bar-mode remapping.
-   */
-  scaleValueToRatio(value) {
-    return this.scale.toRatio(value);
-  }
-
-  /**
-   * Converts a scale value into an absolute angle using the raw scale mapping.
-   *
-   * @param {number} value - Scale value to map.
-   * @returns {number} Absolute arc angle.
-   */
-  scaleValueToAngle(value) {
-    return this.ratioToAngle(this.scaleValueToRatio(value));
-  }
-
-  /**
-   * Converts a scale value into a visual ratio after bar-mode remapping.
-   *
-   * @param {number} value - Scale value to map.
-   * @returns {number} Normalized visual arc position.
-   */
-  valueToRatio(value) {
-    return this.valueMapper.valueToRatio(value);
-  }
-
-  /**
-   * Converts a scale value into an absolute arc angle.
-   */
-  valueToAngle(value) {
-    return this.ratioToAngle(this.valueToRatio(value));
-  }
-
-  /**
-   * Projects an angle and radius to SVG coordinates.
-   */
-  pointAt(angle, radius) {
-    const rad = angle * DEG_TO_RAD;
-
-    return {
-      x: this.cx + Math.cos(rad) * radius,
-      y: this.cy + Math.sin(rad) * radius,
-    };
   }
 }
