--- src/horseshoe-shapes.js
@@ -1,1455 +1,0 @@
-import Colors from './colors.js';
-import ConfigHelper from './config-helper.js';
-
-/**
- * Returns arrays unchanged so color stop and label builders can iterate predictably.
- *
- * @param {*} value - Value that may already be an array.
- * @returns {Array} The original array, or an empty array.
- */
-const asArray = (value) => (Array.isArray(value) ? value : []);
-
-/**
- * Returns color-stop points in physical arc order for the current geometry.
- * Absolute mode keeps signed source values for color lookup while folding the
- * selected negative or positive branch onto the complete horseshoe.
- */
-function buildActiveColorPoints(runtimeConfig, geometry) {
-  const colorStops = asArray(runtimeConfig.colorstops?.colors);
-  const range = geometry.getActiveSourceRange();
-  const activeStops = geometry.getActiveColorStops(colorStops);
-  const boundaryPoints = [
-    {
-      value: range.start,
-      color: Colors.calculateStrokeColor(range.start, runtimeConfig.colorstops, false),
-    },
-    ...activeStops,
-    {
-      value: range.end,
-      color: Colors.calculateStrokeColor(range.end, runtimeConfig.colorstops, false),
-    },
-  ];
-
-  return boundaryPoints
-    .sort((pointA, pointB) => geometry.valueToRatio(pointA.value) - geometry.valueToRatio(pointB.value))
-    .filter((point, index, points) => points.findIndex((candidate) => Number(candidate.value) === Number(point.value)) === index);
-}
-
-/**
- * Builds closed SVG band paths for arcs with butt or round caps.
- */
-class ArcPathBuilder {
-  /**
-   * Converts arc angles and band dimensions into a closed SVG path.
-   *
-   * @param {object} options - Path build options.
-   * @param {GaugeGeometry} options.geometry - Geometry helper for point projection.
-   * @param {object} options.arc - Arc angles, cap style, and visibility.
-   * @param {object} options.band - Band radius and width.
-   * @returns {string} SVG path data for the closed arc band.
-   */
-  static buildBandPath(options = {}) {
-    const { geometry, arc = {}, band = {} } = options;
-
-    if (!geometry || arc.visible === false) {
-      return '';
-    }
-
-    const bandArc = {
-      startAngle: 0,
-      endAngle: 0,
-      startCap: 'butt',
-      endCap: 'butt',
-      ...arc,
-    };
-
-    const arcBand = {
-      radius: geometry.radius,
-      width: 1,
-      ...band,
-    };
-
-    const startAngle = Number(bandArc.startAngle);
-    const endAngle = Number(bandArc.endAngle);
-    const radius = Number(arcBand.radius);
-    const width = Number(arcBand.width);
-
-    if (!Number.isFinite(startAngle) || !Number.isFinite(endAngle) || !Number.isFinite(radius) || !Number.isFinite(width)) {
-      return '';
-    }
-
-    if (endAngle === startAngle || width <= 0) {
-      return '';
-    }
-
-    const innerRadius = radius - width / 2;
-    const outerRadius = radius + width / 2;
-
-    if (innerRadius <= 0 || outerRadius <= 0) {
-      return '';
-    }
-
-    // Project both edges of the band so the path can close as a filled shape.
-    const outerStart = geometry.pointAt(startAngle, outerRadius);
-    const outerEnd = geometry.pointAt(endAngle, outerRadius);
-    const innerEnd = geometry.pointAt(endAngle, innerRadius);
-    const innerStart = geometry.pointAt(startAngle, innerRadius);
-
-    const largeArc = Math.abs(endAngle - startAngle) > 180 ? 1 : 0;
-    const sweep = endAngle > startAngle ? 1 : 0;
-    const reverseSweep = sweep ? 0 : 1;
-    const capRadius = width / 2;
-
-    const parts = [];
-
-    // Draw the outer arc, connect to the inner arc, then close back to the start.
-    parts.push(`M ${outerStart.x} ${outerStart.y}`);
-    parts.push(`A ${outerRadius} ${outerRadius} 0 ${largeArc} ${sweep} ${outerEnd.x} ${outerEnd.y}`);
-
-    if (bandArc.endCap === 'round') {
-      parts.push(`A ${capRadius} ${capRadius} 0 0 ${sweep} ${innerEnd.x} ${innerEnd.y}`);
-    } else {
-      parts.push(`L ${innerEnd.x} ${innerEnd.y}`);
-    }
-
-    parts.push(`A ${innerRadius} ${innerRadius} 0 ${largeArc} ${reverseSweep} ${innerStart.x} ${innerStart.y}`);
-
-    if (bandArc.startCap === 'round') {
-      parts.push(`A ${capRadius} ${capRadius} 0 0 ${sweep} ${outerStart.x} ${outerStart.y}`);
-    } else {
-      parts.push(`L ${outerStart.x} ${outerStart.y}`);
-    }
-
-    parts.push('Z');
-
-    return parts.join(' ');
-  }
-}
-
-/**
- * Builds scale arc segments from configured color stops and applies segment gaps.
- *
- * @param {object} runtimeConfig - Normalized horseshoe runtime configuration.
- * @param {GaugeGeometry} geometry - Geometry helper used to map values to angles.
- * @returns {Array<object>} Scale arc definitions for the renderer.
- */
-function buildColorStopScaleArcs(runtimeConfig, geometry) {
-  const colorStops = asArray(runtimeConfig.colorstops?.colors);
-  const gap = Number(runtimeConfig.colorstops?.gap ?? 0);
-  const scaleArcs = [];
-
-  if (!colorStops.length) {
-    return [
-      {
-        key: 'scale',
-        startAngle: geometry.startAngle,
-        endAngle: geometry.endAngle,
-        startCap: runtimeConfig.horseshoe_scale.linecap.start,
-        endCap: runtimeConfig.horseshoe_scale.linecap.end,
-        color: runtimeConfig.horseshoe_scale.color,
-      },
-    ];
-  }
-
-  const scalePoints = buildActiveColorPoints(runtimeConfig, geometry);
-
-  for (let i = 0; i < scalePoints.length - 1; i += 1) {
-    const pointA = scalePoints[i];
-    const pointB = scalePoints[i + 1];
-    const colorStopStartAngle = geometry.valueToAngle(pointA.value);
-    const colorStopEndAngle = geometry.valueToAngle(pointB.value);
-    const isFirst = i === 0;
-    const isLast = i === scalePoints.length - 2;
-    const drawStartAngle = isFirst ? colorStopStartAngle : colorStopStartAngle + gap / 2;
-    const drawEndAngle = isLast ? colorStopEndAngle : colorStopEndAngle - gap / 2;
-    const visible = drawEndAngle > drawStartAngle;
-
-    scaleArcs.push({
-      key: `scale-colorstop-${i}`,
-      startAngle: visible ? drawStartAngle : 0,
-      endAngle: visible ? drawEndAngle : 0,
-      startCap: 'butt',
-      endCap: 'butt',
-      color: pointA.color,
-      value: pointA.value,
-      visible,
-    });
-  }
-
-  const visibleScaleArcs = scaleArcs.filter((arc) => arc.visible !== false);
-
-  if (visibleScaleArcs.length) {
-    visibleScaleArcs[0].startCap = runtimeConfig.horseshoe_scale.linecap.start;
-    visibleScaleArcs[visibleScaleArcs.length - 1].endCap = runtimeConfig.horseshoe_scale.linecap.end;
-  }
-
-  return scaleArcs;
-}
-
-/**
- * Builds the visible scale arcs for the selected scale display mode.
- *
- * @param {object} runtimeConfig - Normalized horseshoe runtime configuration.
- * @param {GaugeGeometry} geometry - Geometry helper for scale angles.
- * @returns {Array<object>} Scale arc definitions.
- */
-export function buildScaleArcs(runtimeConfig, geometry) {
-  const scaleMode = runtimeConfig.show?.scale_style ?? 'fixed';
-
-  if (scaleMode === 'none') {
-    return [];
-  }
-
-  if (scaleMode === 'colorstopsegments') {
-    return buildColorStopScaleArcs(runtimeConfig, geometry);
-  }
-
-  return [
-    {
-      key: 'scale',
-      startAngle: geometry.startAngle,
-      endAngle: geometry.endAngle,
-      startCap: runtimeConfig.horseshoe_scale.linecap.start,
-      endCap: runtimeConfig.horseshoe_scale.linecap.end,
-      color: runtimeConfig.horseshoe_scale.color,
-    },
-  ];
-}
-
-/**
- * Builds fixed or color-stop arc background items for horseshoe-related layers.
- *
- * @param {object} runtimeConfig - Normalized horseshoe runtime configuration.
- * @param {GaugeGeometry} geometry - Geometry helper for background arcs.
- * @param {object} options - Background mode, dimensions, and key metadata.
- * @returns {Array<object>} Background arc items.
- */
-export function buildArcBackgroundItems(runtimeConfig, geometry, options = {}) {
-  const {
-    mode = 'none',
-    config = {},
-    radius = geometry.radius,
-    width = 6,
-    gap = 0,
-    keyPrefix = 'background',
-  } = options;
-
-  if (mode === 'none') {
-    return [];
-  }
-
-  if (mode === 'lineargradient' || mode === 'colorstopgradient') {
-    const lineCap = config.linecap ?? 'round';
-
-    // Function declaration is intentionally hoisted so all background mode selection stays together.
-    // eslint-disable-next-line no-use-before-define
-    return buildGradientArcPathItems(runtimeConfig, geometry, {
-      mode,
-      radius,
-      width,
-      startCap: lineCap,
-      endCap: lineCap,
-      keyPrefix,
-    });
-  }
-
-  if (mode === 'colorstopsegments') {
-    const colorStops = asArray(runtimeConfig.colorstops?.colors);
-
-    if (!colorStops.length) {
-      return [];
-    }
-
-    const backgroundItems = [];
-    const segmentPoints = buildActiveColorPoints(runtimeConfig, geometry);
-
-    for (let i = 0; i < segmentPoints.length - 1; i += 1) {
-      const pointA = segmentPoints[i];
-      const pointB = segmentPoints[i + 1];
-      const startAngle = geometry.valueToAngle(pointA.value);
-      const endAngle = geometry.valueToAngle(pointB.value);
-      const isFirst = i === 0;
-      const isLast = i === segmentPoints.length - 2;
-      // Preserve the outside caps while applying gaps only between internal segments.
-      const drawStartAngle = isFirst ? startAngle : startAngle + gap / 2;
-      const drawEndAngle = isLast ? endAngle : endAngle - gap / 2;
-
-      if (drawEndAngle > drawStartAngle) {
-        const lineCap = config.linecap ?? 'round';
-        const arc = {
-          key: `${keyPrefix}-colorstop-${i}`,
-          startAngle: drawStartAngle,
-          endAngle: drawEndAngle,
-          startCap: isFirst ? lineCap : 'butt',
-          endCap: isLast ? lineCap : 'butt',
-        };
-
-        backgroundItems.push({
-          key: arc.key,
-          arc,
-          path: ArcPathBuilder.buildBandPath({
-            geometry,
-            arc,
-            band: { radius, width },
-          }),
-          startAngle: drawStartAngle,
-          endAngle: drawEndAngle,
-          radius,
-          width,
-          color: pointA.color,
-          lineCap,
-        });
-      }
-    }
-
-    return backgroundItems;
-  }
-
-  if (mode === 'fixed') {
-    const lineCap = config.linecap ?? 'round';
-    const arc = {
-      key: `${keyPrefix}-fixed`,
-      startAngle: geometry.startAngle,
-      endAngle: geometry.endAngle,
-      startCap: lineCap,
-      endCap: lineCap,
-    };
-
-    return [
-      {
-        key: arc.key,
-        arc,
-        path: ArcPathBuilder.buildBandPath({
-          geometry,
-          arc,
-          band: { radius, width },
-        }),
-        startAngle: geometry.startAngle,
-        endAngle: geometry.endAngle,
-        radius,
-        width,
-        color: config.color,
-        lineCap,
-      },
-    ];
-  }
-
-  return [];
-}
-
-const COLORSTOP_GRADIENT_MAX_ARC_LENGTH = 6;
-
-/**
- * Builds full-band gradient paths for scale, state, and background layers.
- *
- * A colorstopgradient keeps numeric stops and the configured scale mapping.
- * A lineargradient uses normalized positions so every color receives an equal
- * part of the visible arc. Both modes then use the same adaptive path geometry.
- *
- * @param {object} runtimeConfig - Normalized horseshoe runtime configuration.
- * @param {GaugeGeometry} geometry - Geometry helper for value-to-angle mapping.
- * @param {object} options - Gradient mode, band dimensions, caps, and key prefix.
- * @returns {Array<object>} Static full-scale gradient path items.
- */
-function buildGradientArcPathItems(runtimeConfig, geometry, options = {}) {
-  if (runtimeConfig.bar_mode === 'absolute') {
-    return buildAbsoluteGradientArcPathItems(runtimeConfig, geometry, options);
-  }
-  const {
-    mode = 'colorstopgradient',
-    radius = geometry.radius,
-    width = runtimeConfig.horseshoe_state.width,
-    startCap = runtimeConfig.horseshoe_state.linecap.start,
-    endCap = runtimeConfig.horseshoe_state.linecap.end,
-    keyPrefix = mode,
-  } = options;
-  const min = Number(runtimeConfig.horseshoe_scale.min);
-  const max = Number(runtimeConfig.horseshoe_scale.max);
-
-  // The style name determines the coordinate system. lineargradient promises
-  // an even visual color distribution, while colorstopgradient promises that
-  // colors stay attached to their numeric scale and spline positions.
-  const gradientMin = mode === 'lineargradient' ? 0 : min;
-  const gradientMax = mode === 'lineargradient' ? 1 : max;
-  const gradientValueToAngle = mode === 'lineargradient'
-    ? (gradientValue) => geometry.startAngle + gradientValue * geometry.arcDegrees
-    : (gradientValue) => geometry.valueToAngle(gradientValue);
-  const configuredColorStops = asArray(runtimeConfig.colorstops.colors);
-
-  // Only lineargradient replaces numeric stop positions. It preserves the
-  // configured color order but gives every color an equal share of 0..1.
-  // colorstopgradient reuses the original stops without modification.
-  const gradientColorStops = mode === 'lineargradient'
-    ? {
-        colors: configuredColorStops.map((colorStop, index) => ({
-          value: index / (configuredColorStops.length - 1),
-          color: colorStop.color,
-        })),
-      }
-    : runtimeConfig.colorstops;
-  const colorStops = gradientColorStops.colors;
-
-  // Boundary anchors make the selected style cover the complete band even
-  // when a colorstopgradient template omits an explicit min or max stop.
-  const anchorValues = [
-    gradientMin,
-    gradientMax,
-    ...colorStops.map((stop) => Number(stop.value)),
-  ];
-  const sortedAnchors = [...new Set(anchorValues)]
-    .filter((anchorValue) => anchorValue >= gradientMin && anchorValue <= gradientMax)
-    .sort((valueA, valueB) => valueA - valueB);
-  const gradientArcs = [];
-
-  // Color-stop intervals split only when their visible arc length requires a closer curve approximation.
-  for (let anchorIndex = 0; anchorIndex < sortedAnchors.length - 1; anchorIndex += 1) {
-    const intervals = [{
-      startValue: sortedAnchors[anchorIndex],
-      endValue: sortedAnchors[anchorIndex + 1],
-    }];
-
-    while (intervals.length) {
-      const interval = intervals.pop();
-      const startAngle = gradientValueToAngle(interval.startValue);
-      const endAngle = gradientValueToAngle(interval.endValue);
-      const arcLength = radius * Math.abs(endAngle - startAngle) * Math.PI / 180;
-
-      // Split by visible arc length, not by a fixed segment count. Large arcs
-      // stay smooth while spline-compressed intervals do not create excess SVG.
-      if (arcLength > COLORSTOP_GRADIENT_MAX_ARC_LENGTH) {
-        const midValue = (interval.startValue + interval.endValue) / 2;
-
-        // Push right before left so stack processing preserves ascending scale order.
-        intervals.push({ startValue: midValue, endValue: interval.endValue });
-        intervals.push({ startValue: interval.startValue, endValue: midValue });
-      } else {
-        gradientArcs.push({
-          startValue: interval.startValue,
-          endValue: interval.endValue,
-          startAngle,
-          endAngle,
-        });
-      }
-    }
-  }
-
-  return gradientArcs.map((gradientArc, index) => {
-    const isLast = index === gradientArcs.length - 1;
-    const direction = gradientArc.endAngle >= gradientArc.startAngle ? 1 : -1;
-    const segmentDegrees = Math.abs(gradientArc.endAngle - gradientArc.startAngle);
-
-    // Each path receives a gradient in its own arc direction. The 50% overlap
-    // removes rasterization gaps; matching edge colors keep the join invisible.
-    const overlapDegrees = segmentDegrees * 0.5;
-    const gradientStart = geometry.pointAt(gradientArc.startAngle, radius);
-    const gradientEnd = geometry.pointAt(gradientArc.endAngle, radius);
-    const arc = {
-      key: `${keyPrefix}-${index}`,
-      startAngle: gradientArc.startAngle,
-      endAngle: isLast ? gradientArc.endAngle : gradientArc.endAngle + direction * overlapDegrees,
-      startCap: index === 0 ? startCap : 'butt',
-      endCap: isLast ? endCap : 'butt',
-      startValue: gradientArc.startValue,
-      endValue: gradientArc.endValue,
-      gradient: {
-        x1: gradientStart.x,
-        y1: gradientStart.y,
-        x2: gradientEnd.x,
-        y2: gradientEnd.y,
-        startColor: Colors.calculateStrokeColor(gradientArc.startValue, gradientColorStops, true),
-        endColor: Colors.calculateStrokeColor(gradientArc.endValue, gradientColorStops, true),
-      },
-    };
-
-    return {
-      key: arc.key,
-      arc,
-      path: ArcPathBuilder.buildBandPath({
-        geometry,
-        arc,
-        band: {
-          radius,
-          width,
-        },
-      }),
-    };
-  });
-}
-
-function buildAbsoluteGradientArcPathItems(runtimeConfig, geometry, options = {}) {
-  const {
-    mode = 'colorstopgradient',
-    radius = geometry.radius,
-    width = runtimeConfig.horseshoe_state.width,
-    startCap = runtimeConfig.horseshoe_state.linecap.start,
-    endCap = runtimeConfig.horseshoe_state.linecap.end,
-    keyPrefix = mode,
-  } = options;
-  const activePoints = buildActiveColorPoints(runtimeConfig, geometry);
-  const gradientPoints = mode === 'lineargradient'
-    ? activePoints.map((point, index) => ({
-        value: index / (activePoints.length - 1),
-        sourceValue: point.value,
-        color: point.color,
-      }))
-    : activePoints.map((point) => ({
-        value: point.value,
-        sourceValue: point.value,
-        color: point.color,
-      }));
-  const valueToAngle = mode === 'lineargradient'
-    ? (value) => geometry.startAngle + Number(value) * geometry.arcDegrees
-    : (value) => geometry.valueToAngle(value);
-  const gradientColorStops = {
-    // Negative absolute paths run from zero toward decreasing source values.
-    // Color interpolation still requires its lookup stops in numeric order.
-    colors: gradientPoints
-      .map((point) => ({
-        value: point.value,
-        color: point.color,
-      }))
-      .sort((pointA, pointB) => pointA.value - pointB.value),
-  };
-  const gradientArcs = [];
-
-  for (let pointIndex = 0; pointIndex < gradientPoints.length - 1; pointIndex += 1) {
-    const pointA = gradientPoints[pointIndex];
-    const pointB = gradientPoints[pointIndex + 1];
-    const intervals = [{
-      startValue: pointA.value,
-      endValue: pointB.value,
-    }];
-
-    while (intervals.length) {
-      const interval = intervals.pop();
-      const startAngle = valueToAngle(interval.startValue);
-      const endAngle = valueToAngle(interval.endValue);
-      const arcLength = radius * Math.abs(endAngle - startAngle) * Math.PI / 180;
-
-      if (arcLength > COLORSTOP_GRADIENT_MAX_ARC_LENGTH) {
-        const midValue = (interval.startValue + interval.endValue) / 2;
-        intervals.push({ startValue: midValue, endValue: interval.endValue });
-        intervals.push({ startValue: interval.startValue, endValue: midValue });
-      } else {
-        gradientArcs.push({
-          startValue: interval.startValue,
-          endValue: interval.endValue,
-          startAngle,
-          endAngle,
-        });
-      }
-    }
-  }
-
-  return gradientArcs.map((gradientArc, index) => {
-    const isLast = index === gradientArcs.length - 1;
-    const segmentDegrees = Math.abs(gradientArc.endAngle - gradientArc.startAngle);
-    const overlapDegrees = segmentDegrees * 0.5;
-    const gradientStart = geometry.pointAt(gradientArc.startAngle, radius);
-    const gradientEnd = geometry.pointAt(gradientArc.endAngle, radius);
-    const arc = {
-      key: `${keyPrefix}-${index}`,
-      startAngle: gradientArc.startAngle,
-      endAngle: isLast ? gradientArc.endAngle : gradientArc.endAngle + overlapDegrees,
-      startCap: index === 0 ? startCap : 'butt',
-      endCap: isLast ? endCap : 'butt',
-      startValue: gradientArc.startValue,
-      endValue: gradientArc.endValue,
-      gradient: {
-        x1: gradientStart.x,
-        y1: gradientStart.y,
-        x2: gradientEnd.x,
-        y2: gradientEnd.y,
-        startColor: Colors.calculateStrokeColor(gradientArc.startValue, gradientColorStops, true),
-        endColor: Colors.calculateStrokeColor(gradientArc.endValue, gradientColorStops, true),
-      },
-    };
-
-    return {
-      key: arc.key,
-      arc,
-      path: ArcPathBuilder.buildBandPath({
-        geometry,
-        arc,
-        band: { radius, width },
-      }),
-    };
-  });
-}
-
-/**
- * Builds the cached full-scale micro-arcs revealed by the state clip.
- *
- * @param {object} runtimeConfig - Normalized horseshoe runtime configuration.
- * @param {GaugeGeometry} geometry - Geometry helper for value-to-angle mapping.
- * @returns {Array<object>} Static full-scale gradient path items.
- */
-export function buildColorStopGradientPathItems(runtimeConfig, geometry) {
-  return buildGradientArcPathItems(runtimeConfig, geometry);
-}
-
-/**
- * Builds state arc segments clipped to the active value range and color stops.
- *
- * @param {object} runtimeConfig - Normalized horseshoe runtime configuration.
- * @param {GaugeGeometry} geometry - Geometry helper used to map color stop values.
- * @param {number} fromAngle - Start angle of the active state range.
- * @param {number} toAngle - End angle of the active state range.
- * @returns {Array<object>} State arc definitions for color stop segments.
- */
-function buildColorStopStateArcs(runtimeConfig, geometry, fromAngle, toAngle) {
-  const colorStops = asArray(runtimeConfig.colorstops?.colors);
-  const gap = Number(runtimeConfig.colorstops?.gap ?? 0);
-  const stateArcs = [];
-
-  if (colorStops.length < 2) {
-    return [
-      {
-        key: 'state-value',
-        startAngle: fromAngle,
-        endAngle: toAngle,
-        startCap: runtimeConfig.horseshoe_state.linecap.start,
-        endCap: runtimeConfig.horseshoe_state.linecap.end,
-      },
-    ];
-  }
-
-  const statePoints = buildActiveColorPoints(runtimeConfig, geometry);
-
-  for (let i = 0; i < statePoints.length - 1; i += 1) {
-    const pointA = statePoints[i];
-    const pointB = statePoints[i + 1];
-    const segmentStartAngle = geometry.valueToAngle(pointA.value);
-    const segmentEndAngle = geometry.valueToAngle(pointB.value);
-    const clippedStartAngle = Math.max(segmentStartAngle, fromAngle);
-    const clippedEndAngle = Math.min(segmentEndAngle, toAngle);
-    const drawStartAngle = clippedStartAngle === fromAngle ? clippedStartAngle : clippedStartAngle + gap / 2;
-    const drawEndAngle = clippedEndAngle === toAngle ? clippedEndAngle : clippedEndAngle - gap / 2;
-    const visible = drawEndAngle > drawStartAngle;
-
-    stateArcs.push({
-      key: `colorstop-${i}`,
-      startAngle: visible ? drawStartAngle : 0,
-      endAngle: visible ? drawEndAngle : 0,
-      startCap: 'butt',
-      endCap: 'butt',
-      color: pointA.color,
-      value: pointA.value,
-      label: pointA.label,
-      visible,
-    });
-  }
-
-  const visibleStateArcs = stateArcs.filter((arc) => arc.visible !== false);
-
-  if (visibleStateArcs.length) {
-    visibleStateArcs[0].startCap = runtimeConfig.horseshoe_state.linecap.start;
-    visibleStateArcs[visibleStateArcs.length - 1].endCap = runtimeConfig.horseshoe_state.linecap.end;
-  }
-
-  return stateArcs;
-}
-
-/**
- * Selects the state arc color strategy and returns the arc definitions for it.
- *
- * @param {object} runtimeConfig - Normalized horseshoe runtime configuration.
- * @param {GaugeGeometry} geometry - Geometry helper for state angles.
- * @param {number} value - Current numeric state value.
- * @param {object} arcRange - Start and end angle limits for the state arc.
- * @returns {Array<object>} State arc definitions.
- */
-function buildColorAwareStateArcs(runtimeConfig, geometry, value, arcRange) {
-  const strokeStyle = runtimeConfig.show?.horseshoe_style;
-
-  const fromAngle = Number(arcRange.fromAngle ?? geometry.startAngle);
-  const toAngle = Number(arcRange.toAngle ?? geometry.startAngle);
-
-  if (strokeStyle === 'colorstopsegments') {
-    return buildColorStopStateArcs(runtimeConfig, geometry, fromAngle, toAngle);
-  }
-
-  if (strokeStyle === 'autominmax') {
-    const bidirectional = runtimeConfig.bar_mode === 'bidirectional' || runtimeConfig.bar_mode === 'bidirectional_symmetrical' || runtimeConfig.bar_mode === 'bidirectional_linear';
-    const absolute = runtimeConfig.bar_mode === 'absolute';
-    let color = Colors.calculateStrokeColor(value, runtimeConfig.colorstopsMinMax, true);
-
-    // Bidirectional rendering keeps its original min-to-zero and zero-to-max
-    // interpolation because its negative arc still runs toward the zero point.
-    if (bidirectional) {
-      const numericValue = Number(value);
-      const min = Number(runtimeConfig.horseshoe_scale.min);
-      const max = Number(runtimeConfig.horseshoe_scale.max);
-      const colorStops = runtimeConfig.colorstops.colors;
-      const firstColor = colorStops[0].color;
-      const lastColor = colorStops[colorStops.length - 1].color;
-      const zeroColor = Colors.calculateStrokeColor(0, runtimeConfig.colorstops, true);
-
-      color = numericValue < 0
-        ? Colors.getGradientValue(firstColor, zeroColor, (numericValue - min) / (0 - min))
-        : Colors.getGradientValue(zeroColor, lastColor, numericValue / max);
-    }
-
-    // Absolute rendering starts both signed branches at zero and follows the
-    // clamped branch ratio supplied by geometry toward the signed endpoint.
-    if (absolute) {
-      const numericValue = Number(value);
-      const range = geometry.getActiveSourceRange();
-      const zeroColor = Colors.calculateStrokeColor(0, runtimeConfig.colorstops, true);
-      const endColor = Colors.calculateStrokeColor(range.end, runtimeConfig.colorstops, true);
-
-      color = Colors.getGradientValue(zeroColor, endColor, geometry.valueToRatio(numericValue));
-    }
-
-    return [
-      {
-        key: 'state-value',
-        startAngle: fromAngle,
-        endAngle: toAngle,
-        startCap: runtimeConfig.horseshoe_state.linecap.start,
-        endCap: runtimeConfig.horseshoe_state.linecap.end,
-        color,
-      },
-    ];
-  }
-
-  if (strokeStyle === 'minmaxgradient' || strokeStyle === 'lineargradient') {
-    const colorStops = runtimeConfig.colorstops.colors;
-    let startColor = colorStops[0].color;
-    let endColor = colorStops[colorStops.length - 1].color;
-    const bidirectional = runtimeConfig.bar_mode === 'bidirectional' || runtimeConfig.bar_mode === 'bidirectional_symmetrical' || runtimeConfig.bar_mode === 'bidirectional_linear';
-    const absolute = runtimeConfig.bar_mode === 'absolute';
-    let activeColors = colorStops.map((colorStop) => colorStop.color);
-
-    if (bidirectional || absolute) {
-      const zeroColor = Colors.calculateStrokeColor(0, runtimeConfig.colorstops, true);
-
-      if (Number(value) < 0) {
-        const negativeColors = colorStops
-          .filter((colorStop) => Number(colorStop.value) < 0)
-          .map((colorStop) => colorStop.color);
-
-        if (absolute) {
-          startColor = zeroColor;
-          endColor = negativeColors[0];
-          activeColors = [zeroColor, ...negativeColors.reverse()];
-        } else {
-          endColor = zeroColor;
-          activeColors = [...negativeColors, zeroColor];
-        }
-      } else {
-        startColor = zeroColor;
-        activeColors = [
-          zeroColor,
-          ...colorStops.filter((colorStop) => Number(colorStop.value) > 0).map((colorStop) => colorStop.color),
-        ];
-      }
-    }
-
-    // Linear gradients distribute every selected color evenly across the active arc.
-    const linearGradientColorStops = {
-      colors: activeColors.map((colorStop, index) => ({
-        value: index / (activeColors.length - 1),
-        color: colorStop,
-      })),
-    };
-    const direction = toAngle >= fromAngle ? 1 : -1;
-    const activeDegrees = Math.abs(toAngle - fromAngle);
-    const activeArcLength = geometry.radius * activeDegrees * Math.PI / 180;
-    const activeSegmentCount = Math.ceil(activeArcLength / COLORSTOP_GRADIENT_MAX_ARC_LENGTH);
-    const maximumArcLength = geometry.radius * Math.abs(geometry.arcDegrees) * Math.PI / 180;
-    const maximumSegmentCount = Math.ceil(maximumArcLength / COLORSTOP_GRADIENT_MAX_ARC_LENGTH);
-    const stateArcs = [];
-
-    // Keep a stable maximum path pool for direct animation updates, but draw only the adaptive active segment count.
-    for (let index = 0; index < maximumSegmentCount; index += 1) {
-      const visible = index < activeSegmentCount;
-      const startRatio = visible ? index / activeSegmentCount : 0;
-      const endRatio = visible ? (index + 1) / activeSegmentCount : 0;
-      const segmentStartAngle = fromAngle + direction * activeDegrees * startRatio;
-      const segmentEndAngle = fromAngle + direction * activeDegrees * endRatio;
-      const gradientStart = geometry.pointAt(segmentStartAngle, geometry.radius);
-      const gradientEnd = geometry.pointAt(segmentEndAngle, geometry.radius);
-      const isLast = index === activeSegmentCount - 1;
-      const segmentDegrees = Math.abs(segmentEndAngle - segmentStartAngle);
-      const overlapDegrees = segmentDegrees * 0.5;
-
-      stateArcs.push({
-        key: `${strokeStyle}-${index}`,
-        startAngle: segmentStartAngle,
-        endAngle: visible && !isLast ? segmentEndAngle + direction * overlapDegrees : segmentEndAngle,
-        startCap: index === 0 ? runtimeConfig.horseshoe_state.linecap.start : 'butt',
-        endCap: isLast ? runtimeConfig.horseshoe_state.linecap.end : 'butt',
-        visible,
-        gradient: {
-          x1: gradientStart.x,
-          y1: gradientStart.y,
-          x2: gradientEnd.x,
-          y2: gradientEnd.y,
-          startColor: strokeStyle === 'lineargradient'
-            ? Colors.calculateStrokeColor(startRatio, linearGradientColorStops, true)
-            : startRatio === 0 ? startColor : Colors.getGradientValue(startColor, endColor, startRatio),
-          endColor: strokeStyle === 'lineargradient'
-            ? Colors.calculateStrokeColor(endRatio, linearGradientColorStops, true)
-            : endRatio === 0 ? startColor : endRatio === 1 ? endColor : Colors.getGradientValue(startColor, endColor, endRatio),
-        },
-      });
-    }
-
-    return stateArcs;
-  }
-
-  if (strokeStyle === 'colorstop' || strokeStyle === 'colorstopinterpolated') {
-    return [
-      {
-        key: 'state-value',
-        startAngle: fromAngle,
-        endAngle: toAngle,
-        startCap: runtimeConfig.horseshoe_state.linecap.start,
-        endCap: runtimeConfig.horseshoe_state.linecap.end,
-        color: Colors.calculateStrokeColor(value, runtimeConfig.colorstops, strokeStyle === 'colorstopinterpolated'),
-      },
-    ];
-  }
-
-  if (strokeStyle === 'colorstopgradient') {
-    return [
-      {
-        key: 'colorstopgradient-clip',
-        startAngle: fromAngle,
-        endAngle: toAngle,
-        startCap: runtimeConfig.horseshoe_state.linecap.start,
-        endCap: runtimeConfig.horseshoe_state.linecap.end,
-      },
-    ];
-  }
-
-  return [
-    {
-      key: 'state-value',
-      startAngle: fromAngle,
-      endAngle: toAngle,
-      startCap: runtimeConfig.horseshoe_state.linecap.start,
-      endCap: runtimeConfig.horseshoe_state.linecap.end,
-    },
-  ];
-}
-
-/**
- * Builds the normal unidirectional state arc from the scale start to the value.
- *
- * @param {object} runtimeConfig - Normalized horseshoe runtime configuration.
- * @param {GaugeGeometry} geometry - Geometry helper for value-to-angle mapping.
- * @param {number} value - Current numeric state value.
- * @returns {Array<object>} State arc definitions.
- */
-function buildNormalStateArcs(runtimeConfig, geometry, value) {
-  return buildColorAwareStateArcs(runtimeConfig, geometry, value, {
-    fromAngle: geometry.startAngle,
-    toAngle: geometry.valueToAngle(value),
-  });
-}
-
-/**
- * Builds the bidirectional state arc between zero and the current value.
- *
- * @param {object} runtimeConfig - Normalized horseshoe runtime configuration.
- * @param {GaugeGeometry} geometry - Geometry helper containing the zero angle.
- * @param {number} value - Current numeric state value.
- * @returns {Array<object>} State arc definitions.
- */
-function buildBidirectionalStateArcs(runtimeConfig, geometry, value) {
-  const valueAngle = geometry.valueToAngle(value);
-  const zeroAngle = geometry.zeroAngle;
-
-  return buildColorAwareStateArcs(runtimeConfig, geometry, value, {
-    fromAngle: Math.min(zeroAngle, valueAngle),
-    toAngle: Math.max(zeroAngle, valueAngle),
-  });
-}
-
-/**
- * Compares one mapped state index with the current mapped state index.
- *
- * @param {number} index - Index of the state map entry being processed.
- * @param {number} currentIndex - Index of the current mapped state.
- * @returns {string} Stable before/current/after relation for string-state rendering and labels.
- */
-function getMappedStateRelation(index, currentIndex) {
-  if (currentIndex < 0) {
-    return 'after';
-  }
-
-  if (index < currentIndex) {
-    return 'before';
-  }
-
-  if (index > currentIndex) {
-    return 'after';
-  }
-
-  return 'current';
-}
-
-/**
- * Finds label-specific state-map metadata for one raw state.
- *
- * @param {object} runtimeConfig - Normalized horseshoe runtime configuration.
- * @param {string|number} state - Raw state from the value state map.
- * @returns {object|undefined} Label state-map entry.
- */
-function getStringstateLabelConfig(runtimeConfig) {
-  const mode = runtimeConfig.horseshoe_state?.mode;
-
-  return mode === 'stringstate_mode' || mode === 'stringstate_level'
-    ? runtimeConfig.horseshoe_labels?.[mode]
-    : undefined;
-}
-
-function getLabelStateMapEntry(runtimeConfig, state) {
-  return asArray(getStringstateLabelConfig(runtimeConfig)?.state_map?.map).find((entry) => String(entry.state) === String(state));
-}
-
-function getStringstateRoleStyles(runtimeConfig, relation) {
-  return ConfigHelper.toStyleDict(getStringstateLabelConfig(runtimeConfig)?.[relation]?.styles);
-}
-
-function getStringstateStateRoleStyles(labelStateEntry, relation) {
-  return ConfigHelper.toStyleDict(labelStateEntry?.[relation]?.styles);
-}
-
-/**
- * Builds one segment per mapped state and marks the active segment.
- *
- * @param {object} runtimeConfig - Normalized horseshoe runtime configuration.
- * @param {GaugeGeometry} geometry - Geometry helper that provides the arc size.
- * @param {number} value - Current mapped numeric state value.
- * @returns {Array<object>} Segment arc definitions.
- */
-function buildMappedStateArcs(runtimeConfig, geometry, value) {
-  const stateMap = runtimeConfig.state_map.map;
-  const segmentGap = runtimeConfig.horseshoe_state.segment_gap;
-  const count = stateMap.length;
-
-  if (!count) {
-    return [];
-  }
-
-  // Mapped states occupy equal visual slots. Their value is the segment index.
-  const currentIndex = stateMap.findIndex((item) => Number(item.value) === Number(value));
-  const step = geometry.arcDegrees / count;
-
-  const arcs = stateMap.map((item, index) => {
-    const relation = getMappedStateRelation(index, currentIndex);
-    const active = relation === 'current';
-    const highlighted = runtimeConfig.horseshoe_state.mode === 'stringstate_level'
-      ? relation === 'before' || relation === 'current'
-      : active;
-    const itemValue = Number(item.value);
-    let startAngle = geometry.startAngle + index * step + segmentGap / 2;
-    let endAngle = geometry.startAngle + (index + 1) * step - segmentGap / 2;
-    let segmentColor = item.color ?? Colors.calculateStrokeColor(item.value, runtimeConfig.colorstops, runtimeConfig.show?.horseshoe_style === 'colorstopinterpolated');
-    let segmentStartValue;
-    let segmentEndValue;
-
-
-
-    return {
-      key: `mapped-state-${index}`,
-      startAngle,
-      endAngle,
-      startCap: index === 0 ? runtimeConfig.horseshoe_state.linecap.start : 'butt',
-      endCap: index === count - 1 ? runtimeConfig.horseshoe_state.linecap.end : 'butt',
-      active: highlighted,
-      relation,
-      value: item.value,
-      startValue: segmentStartValue,
-      endValue: segmentEndValue,
-      color: segmentColor,
-      label: item.display_label ?? item.label ?? String(item.state ?? item.value),
-    };
-  });
-
-  return arcs;
-}
-
-/**
- * Chooses the state arc builder for segment, bidirectional, or normal mode.
- *
- * @param {object} runtimeConfig - Normalized horseshoe runtime configuration.
- * @param {GaugeGeometry} geometry - Geometry helper for arc calculations.
- * @param {number} value - Current numeric state value.
- * @returns {Array<object>} State arc definitions.
- */
-function buildStateArcs(runtimeConfig, geometry, value) {
-  if (runtimeConfig.horseshoe_state.mode === 'segment' || runtimeConfig.horseshoe_state.mode === 'stringstate_mode' || runtimeConfig.horseshoe_state.mode === 'stringstate_level') {
-    return buildMappedStateArcs(runtimeConfig, geometry, value);
-  }
-
-  if (runtimeConfig.bar_mode === 'bidirectional' || runtimeConfig.bar_mode === 'bidirectional_symmetrical' || runtimeConfig.bar_mode === 'bidirectional_linear') {
-    return buildBidirectionalStateArcs(runtimeConfig, geometry, value);
-  }
-
-  return buildNormalStateArcs(runtimeConfig, geometry, value);
-}
-
-/**
- * Converts state arc definitions into renderable SVG path items.
- *
- * @param {object} runtimeConfig - Normalized horseshoe runtime configuration.
- * @param {GaugeGeometry} geometry - Geometry helper used by the path builder.
- * @param {number} value - Current numeric state value.
- * @returns {Array<object>} Renderable state path items.
- */
-export function buildStatePathItems(runtimeConfig, geometry, value) {
-  const stateArcs = buildStateArcs(runtimeConfig, geometry, value);
-
-  const stateBand = {
-    radius: geometry.radius,
-    width: runtimeConfig.horseshoe_state.width,
-  };
-
-  return stateArcs.map((arc, index) => ({
-    key: arc.key ?? `state-arc-${index}`,
-    arc,
-    path: ArcPathBuilder.buildBandPath({
-      geometry,
-      arc,
-      band: stateBand,
-    }),
-  }));
-}
-
-/**
- * Converts scale arc definitions into renderable SVG path items.
- *
- * @param {object} runtimeConfig - Normalized horseshoe runtime configuration.
- * @param {GaugeGeometry} geometry - Geometry helper used by the path builder.
- * @returns {Array<object>} Renderable scale path items.
- */
-export function buildScalePathItems(runtimeConfig, geometry) {
-  const scaleMode = runtimeConfig.show?.scale_style ?? 'fixed';
-
-  if (scaleMode === 'lineargradient' || scaleMode === 'colorstopgradient') {
-    return buildGradientArcPathItems(runtimeConfig, geometry, {
-      mode: scaleMode,
-      radius: geometry.radius,
-      width: runtimeConfig.horseshoe_scale.width,
-      startCap: runtimeConfig.horseshoe_scale.linecap.start,
-      endCap: runtimeConfig.horseshoe_scale.linecap.end,
-      keyPrefix: `scale-${scaleMode}`,
-    });
-  }
-
-  const scaleArcs = buildScaleArcs(runtimeConfig, geometry);
-
-  const scaleBand = {
-    radius: geometry.radius,
-    width: runtimeConfig.horseshoe_scale.width,
-  };
-
-  return scaleArcs.map((arc, index) => ({
-    key: arc.key ?? `scale-arc-${index}`,
-    arc,
-    path: ArcPathBuilder.buildBandPath({
-      geometry,
-      arc,
-      band: scaleBand,
-    }),
-  }));
-}
-
-/**
- * Resolves one label stop to its value, angle, radius, and SVG coordinates.
- *
- * @param {object} runtimeConfig - Normalized horseshoe runtime configuration.
- * @param {GaugeGeometry} geometry - Geometry helper for angle and point projection.
- * @param {object} labelConfig - Label stop metadata.
- * @returns {object} Positioned label item for the renderer.
- */
-function buildLabelItem(runtimeConfig, geometry, labelConfig = {}) {
-  const value = Number(labelConfig.value);
-  const angle = labelConfig.angle !== undefined ? Number(labelConfig.angle) : geometry.valueToAngle(value);
-  const startAngle = labelConfig.startValue !== undefined ? geometry.valueToAngle(Number(labelConfig.startValue)) : undefined;
-  const endAngle = labelConfig.endValue !== undefined ? geometry.valueToAngle(Number(labelConfig.endValue)) : undefined;
-  const arcSize = startAngle !== undefined && endAngle !== undefined ? Math.max(1, Math.abs(endAngle - startAngle)) : undefined;
-  const radius = geometry.radius + Number(runtimeConfig.horseshoe_labels.offset ?? runtimeConfig.horseshoe_state.width + 2);
-  const point = geometry.pointAt(angle, radius);
-
-  return {
-    ...labelConfig,
-    value,
-    text: labelConfig.text ?? String(value),
-    role: labelConfig.role ?? 'label',
-    angle,
-    arcSize,
-    radius,
-    x: point.x,
-    y: point.y,
-  };
-}
-
-/**
- * Builds numeric tick values outward from the supplied anchor.
- *
- * @param {number} min - First scale value.
- * @param {number} max - Last scale value.
- * @param {number} ticksize - Distance between generated tick values.
- * @param {number} anchor - Value from which both tick branches are counted.
- * @returns {Array<number>} Numeric tick values.
- */
-function buildTickValues(min, max, ticksize, anchor) {
-  const values = [anchor];
-
-  for (let value = anchor - ticksize; value >= min - 1e-9; value -= ticksize) {
-    values.push(Number(value.toFixed(10)));
-  }
-
-  for (let value = anchor + ticksize; value <= max + 1e-9; value += ticksize) {
-    values.push(Number(value.toFixed(10)));
-  }
-
-  return values.sort((valueA, valueB) => valueA - valueB);
-}
-
-/**
- * Builds magnitude labels for the signed branch currently occupying an
- * absolute horseshoe. Values remain signed internally for geometry and color.
- */
-function buildAbsoluteLabelStopItems(runtimeConfig, geometry) {
-  const labelsAt = runtimeConfig.show.labels_at ?? 'none';
-  const magnitudeMax = geometry.getActiveMagnitudeMax();
-  const activeColorStops = geometry.getActiveColorStops(asArray(runtimeConfig.colorstops?.colors));
-  const toLabelStop = (magnitude, role, extra = {}) => ({
-    value: geometry.magnitudeToSourceValue(magnitude),
-    text: String(magnitude),
-    role,
-    magnitude,
-    ...extra,
-  });
-  let labelStops = [];
-
-  if (labelsAt === 'minmax' || labelsAt === 'minmax0') {
-    labelStops = [
-      toLabelStop(0, 'min'),
-      toLabelStop(magnitudeMax, 'max'),
-    ];
-  }
-
-  if (labelsAt === 'colorstop' || labelsAt === 'colorstops' || labelsAt === 'both') {
-    labelStops = [
-      toLabelStop(0, 'min'),
-      ...activeColorStops.map((stop) => ({
-        value: Number(stop.value),
-        text: stop.label ?? String(Math.abs(Number(stop.value))),
-        role: 'colorstop',
-        magnitude: Math.abs(Number(stop.value)),
-        color: stop.color,
-      })),
-      toLabelStop(magnitudeMax, 'max'),
-    ];
-  }
-
-  if (labelsAt === 'ticks_major' || labelsAt === 'both') {
-    const ticksize = Number(runtimeConfig.horseshoe_tickmarks?.ticks_major?.ticksize);
-
-    if (Number.isFinite(ticksize) && ticksize > 0) {
-      const tickLabels = buildTickValues(0, magnitudeMax, ticksize, 0).map((magnitude, index, values) => (
-        toLabelStop(magnitude, index === 0 ? 'min' : index === values.length - 1 ? 'max' : 'tick-major')
-      ));
-
-      labelStops = labelsAt === 'both' ? [...labelStops, ...tickLabels] : tickLabels;
-    }
-  }
-
-  const uniqueStops = labelStops
-    .sort((stopA, stopB) => stopA.magnitude - stopB.magnitude)
-    .filter((stop, index, stops) => stops.findIndex((candidate) => candidate.magnitude === stop.magnitude) === index);
-  const distanceMin = Number(runtimeConfig.horseshoe_labels.distance_min ?? 0);
-  const visibleStops = [];
-
-  uniqueStops.forEach((stop) => {
-    const previous = visibleStops[visibleStops.length - 1];
-
-    if (!previous || distanceMin <= 0 || stop.magnitude - previous.magnitude >= distanceMin) {
-      visibleStops.push(stop);
-    }
-  });
-
-  if (visibleStops.length) {
-    visibleStops[0].role = 'min';
-    visibleStops[visibleStops.length - 1].role = 'max';
-  }
-
-  return visibleStops;
-}
-
-/**
- * Builds and filters label stops for min/max, color stops, major ticks, or both.
- *
- * @param {object} runtimeConfig - Normalized horseshoe runtime configuration.
- * @returns {Array<object>} Label stop definitions before positioning.
- */
-export function buildLabelStopItems(runtimeConfig, geometry) {
-  const labelsAt = runtimeConfig.show.labels_at ?? 'none';
-
-  if (runtimeConfig.bar_mode === 'absolute' && labelsAt !== 'segment' && labelsAt !== 'stringstate') {
-    return buildAbsoluteLabelStopItems(runtimeConfig, geometry);
-  }
-  const min = Number(runtimeConfig.horseshoe_scale.min);
-  const max = Number(runtimeConfig.horseshoe_scale.max);
-  const colorStops = asArray(runtimeConfig.colorstops?.colors);
-  let labelStops = [];
-
-  if (labelsAt === 'minmax') {
-    labelStops = [
-      { value: min, text: String(min), role: 'min' },
-      { value: max, text: String(max), role: 'max' },
-    ];
-  }
-
-  if (labelsAt === 'minmax0') {
-    labelStops = [
-      { value: min, text: String(min), role: 'min' },
-      { value: 0, text: '0', role: 'zero' },
-      { value: max, text: String(max), role: 'max' },
-    ];
-  }
-
-  if (labelsAt === 'colorstop' || labelsAt === 'colorstops') {
-    labelStops = [
-      { value: min, text: String(min), role: 'min' },
-      ...colorStops.map((stop) => ({
-        value: stop.value,
-        text: stop.label ?? String(stop.value),
-        role: 'colorstop',
-        color: stop.color,
-      })),
-      { value: max, text: String(max), role: 'max' },
-    ];
-  }
-
-  if (labelsAt === 'ticks_major') {
-    const ticksize = Number(runtimeConfig.horseshoe_tickmarks?.ticks_major?.ticksize);
-    const bidirectional = runtimeConfig.bar_mode === 'bidirectional' || runtimeConfig.bar_mode === 'bidirectional_symmetrical' || runtimeConfig.bar_mode === 'bidirectional_linear';
-    const tickAnchor = bidirectional ? 0 : min;
-
-    if (Number.isFinite(ticksize) && ticksize > 0) {
-      labelStops = buildTickValues(min, max, ticksize, tickAnchor).map((value, index, values) => ({
-        value,
-        text: String(value),
-        role: index === 0 ? 'min' : index === values.length - 1 ? 'max' : 'tick-major',
-      }));
-    }
-  }
-
-  if (labelsAt === 'both') {
-    const colorStopLabels = colorStops.length
-      ? [
-          { value: min, text: String(min), role: 'min' },
-          ...colorStops.map((stop) => ({
-            value: stop.value,
-            text: stop.label ?? String(stop.value),
-            role: 'colorstop',
-            color: stop.color,
-          })),
-          { value: max, text: String(max), role: 'max' },
-        ]
-      : [];
-
-    const ticksize = Number(runtimeConfig.horseshoe_tickmarks?.ticks_major?.ticksize);
-    const bidirectional = runtimeConfig.bar_mode === 'bidirectional' || runtimeConfig.bar_mode === 'bidirectional_symmetrical' || runtimeConfig.bar_mode === 'bidirectional_linear';
-    const tickAnchor = bidirectional ? 0 : min;
-    const tickLabels =
-      Number.isFinite(ticksize) && ticksize > 0
-        ? buildTickValues(min, max, ticksize, tickAnchor).map((value) => ({
-            value,
-            text: String(value),
-            role: 'tick-major',
-          }))
-        : [];
-
-    labelStops = [...colorStopLabels, ...tickLabels];
-  }
-
-  if (labelsAt === 'segment' || labelsAt === 'stringstate') {
-    const stateMap = asArray(runtimeConfig.state_map?.map);
-
-    const currentIndex = stateMap.findIndex((item) => Number(item.value) === Number(runtimeConfig.mapped_state?.value));
-
-    if (runtimeConfig.horseshoe_state?.mode === 'stringstate_mode' || runtimeConfig.horseshoe_state?.mode === 'stringstate_level') {
-      labelStops = stateMap.map((item, index) => {
-        const relation = getMappedStateRelation(index, currentIndex);
-        const labelStateEntry = getLabelStateMapEntry(runtimeConfig, item.state);
-        const itemValue = Number(item.value);
-        const labelValue = index + 0.5;
-        const labelStartValue = index;
-        const labelEndValue = index + 1;
-
-        const roleStyles = getStringstateRoleStyles(runtimeConfig, relation);
-        const stateStyles = ConfigHelper.toStyleDict(labelStateEntry?.styles);
-        const stateRoleStyles = getStringstateStateRoleStyles(labelStateEntry, relation);
-        const styles = {
-          ...roleStyles,
-          ...stateStyles,
-          ...stateRoleStyles,
-        };
-
-        if (runtimeConfig.debug_labels || runtimeConfig.dev?.debug_labels) {
-          console.log('[horseshoe-labels] string-state label style', {
-            state: item.state,
-            relation,
-            roleStyles,
-            stateRoleStyles,
-            stateStyles,
-            styles,
-          });
-        }
-
-        return {
-          value: labelValue,
-          startValue: labelStartValue,
-          endValue: labelEndValue,
-          text: labelStateEntry?.label ?? item.display_label ?? String(item.state ?? item.value),
-          role: 'segment',
-          relation,
-          styles,
-        };
-      });
-    } else {
-      labelStops = stateMap.map((item, index) => {
-        const labelStateEntry = getLabelStateMapEntry(runtimeConfig, item.state);
-
-        const relation = getMappedStateRelation(index, currentIndex);
-
-        return {
-          value: item.value,
-          text: labelStateEntry?.label ?? item.display_label ?? String(item.state ?? item.value),
-          role: 'segment',
-          relation,
-          styles: {
-            ...getStringstateRoleStyles(runtimeConfig, relation),
-            ...ConfigHelper.toStyleDict(labelStateEntry?.styles),
-            ...getStringstateStateRoleStyles(labelStateEntry, relation),
-          },
-        };
-      });
-    }
-  }
-
-  // Normalize label stops into sorted, in-range, unique values before applying spacing.
-  const bidirectional = runtimeConfig.bar_mode === 'bidirectional' || runtimeConfig.bar_mode === 'bidirectional_symmetrical' || runtimeConfig.bar_mode === 'bidirectional_linear';
-  const validStops = labelStops
-    .filter((stop) => {
-      const value = Number(stop.value);
-      const inScale = value >= min && value <= max;
-      return Number.isFinite(value) && (inScale || (bidirectional && value === 0));
-    })
-    .sort((a, b) => Number(a.value) - Number(b.value))
-    .filter((stop, index, array) => {
-      const value = Number(stop.value);
-      return array.findIndex((item) => Number(item.value) === value) === index;
-    });
-
-  const distanceMin = Number(runtimeConfig.horseshoe_labels.distance_min ?? 0);
-  const visibleStops = [];
-
-  // distance_min suppresses labels that would be too close in value-space.
-  validStops.forEach((stop) => {
-    const value = Number(stop.value);
-
-    if (distanceMin <= 0) {
-      visibleStops.push(stop);
-      return;
-    }
-
-    const previous = visibleStops[visibleStops.length - 1];
-
-    if (!previous || Math.abs(value - Number(previous.value)) >= distanceMin) {
-      visibleStops.push(stop);
-    }
-  });
-
-  // The first and last visible labels behave as scale boundary labels after filtering.
-  if (visibleStops.length) {
-    visibleStops[0].role = 'min';
-    visibleStops[visibleStops.length - 1].role = 'max';
-  }
-
-  return visibleStops;
-}
-
-/**
- * Builds positioned label items from the configured label stop source.
- *
- * @param {object} runtimeConfig - Normalized horseshoe runtime configuration.
- * @param {GaugeGeometry} geometry - Geometry helper for label positions.
- * @returns {Array<object>} Positioned label items.
- */
-export function buildLabelItems(runtimeConfig, geometry) {
-  return buildLabelStopItems(runtimeConfig, geometry).map((labelStop) => buildLabelItem(runtimeConfig, geometry, labelStop));
-}
-
-/**
- * Builds fixed or color-stop label background arc items.
- *
- * @param {object} runtimeConfig - Normalized horseshoe runtime configuration.
- * @param {GaugeGeometry} geometry - Geometry helper for background arcs.
- * @returns {Array<object>} Label background arc items.
- */
-export function buildLabelBackgroundItems(runtimeConfig, geometry) {
-  const backgroundMode = runtimeConfig.show.label_background ?? 'none';
-  const backgroundConfig = runtimeConfig.horseshoe_labels.background ?? {};
-  const radius = geometry.radius + Number(runtimeConfig.horseshoe_labels.offset ?? runtimeConfig.horseshoe_state.width + 2);
-  const width = Number(backgroundConfig.width ?? 6);
-  const gap = Number(backgroundConfig.gap ?? 0);
-
-  return buildArcBackgroundItems(runtimeConfig, geometry, {
-    mode: backgroundMode,
-    config: backgroundConfig,
-    radius,
-    width,
-    gap,
-    keyPrefix: 'label-background',
-  });
-}
-
-/**
- * Builds the optional horseshoe background arc behind scale and state layers.
- *
- * @param {object} runtimeConfig - Normalized horseshoe runtime configuration.
- * @param {GaugeGeometry} geometry - Geometry helper for background arcs.
- * @returns {Array<object>} Horseshoe background arc items.
- */
-export function buildHorseshoeBackgroundItems(runtimeConfig, geometry) {
-  const backgroundMode = runtimeConfig.show.horseshoe_background ?? 'none';
-  const backgroundConfig = runtimeConfig.horseshoe_background ?? {};
-  const radius = geometry.radius + Number(backgroundConfig.offset ?? 0);
-  const width = Number(backgroundConfig.width ?? runtimeConfig.horseshoe_scale.width ?? runtimeConfig.horseshoe_state.width ?? 6);
-  const gap = Number(backgroundConfig.gap ?? runtimeConfig.colorstops?.gap ?? 0);
-
-  return buildArcBackgroundItems(runtimeConfig, geometry, {
-    mode: backgroundMode,
-    config: backgroundConfig,
-    radius,
-    width,
-    gap,
-    keyPrefix: 'horseshoe-background',
-  });
-}
-
-/**
- * Public wrapper for building a closed SVG band path.
- *
- * @param {GaugeGeometry} geometry - Geometry helper used by the path builder.
- * @param {object} arc - Arc angles, cap style, and visibility.
- * @param {object} band - Band radius and width.
- * @returns {string} SVG path data for the closed arc band.
- */
-export function buildBandPath(geometry, arc, band) {
-  return ArcPathBuilder.buildBandPath({
-    geometry,
-    arc,
-    band,
-  });
-}
